requirements.txt
streamlit>=1.28.0
Pillow>=9.0.0
pandas>=1.5.0
bleach>=6.1.0
html-sanitizer>=2.4.0
python-magic>=0.4.27
validators>=0.22.0
pydantic>=2.0.0
tinycss2>=1.2.0
pypdf>=3.15.0
pdfplumber>=0.9.0
python-docx>=0.8.11
diskcache>=5.6.0
redis>=4.5.0
structlog>=23.1.0
prometheus-client>=0.17.0
psutil>=5.9.0